"Refactor Category module to use CategoryId as string and introduce CategoryPrimitives"
This diff includes changes to the Category module in the devices feature of the frontend application. It refactors the Category module to use CategoryId as a string instead of a number. Additionally, it introduces a new type called CategoryPrimitives, which is used to represent the properties of a Category object in a primitive form. The CategoryGetter module has also been updated to return CategoryPrimitives instead of Category objects. Finally, the CategoryDefaultData type has been updated to use a string as the key instead of a number.