"Update seeders, models, and repositories to use new category and brand data structures"
This commit includes the following changes:

1. Remove the `brands` file from the `deviceData` folder and replace it with a new `BrandData` folder that contains a `brands` file.
2. Update the `keyboardModel` file to use the new `brandsData` instead of the old `brands`.
3. Remove the `ImpFinancieraAlmacen` and `keyboardAlmacen` files from the `deviceData` folder.
4. Update the `SequelizeDeviceRepository` class to use the new `categoryData` and `BrandData` instead of the old `categories` and `brands`.
5. Add a new `CriteriaToSequelizeConverter` class to the `Shared/infrastructure/criteria` folder to handle converting criteria objects to Sequelize find options.
6. Update the `SequelizeDeviceRepository` class to use the new `CriteriaToSequelizeConverter` class.

These changes are part of a larger refactoring effort to improve the organization and maintainability of the codebase. The new `BrandData` and `categoryData` structures will make it easier to manage and update brand and category data in the future. The `CriteriaToSequelizeConverter` class will also make it easier to convert complex criteria objects to Sequelize find options, reducing the amount of boilerplate code in the repositories.