"Change initial filter for Almacen page and update DeviceByCriteriaSearcher and SequelizeDeviceRepository to support 'typeOfSite' filter"
This commit includes changes to the frontend and backend code for the 'Almacen' page. In the frontend, the initial filter for the Almacen page has been changed from statusId to typeOfSite. In the backend, the DeviceByCriteriaSearcher class has been updated to support a new filter operator 'CONTAINS' and the SequelizeDeviceRepository has been updated to include the 'typeOfSite' filter in the query options when the 'typeOfSite' filter is present in the criteria. Additionally, the 'typeOfSite' filter is being logged to the console for debugging purposes.