Change location and site ID types from integers to UUIDs; update corresponding Sequelize models and domain entities
This commit changes the ID types for location and site domain entities from integers to UUIDs. The corresponding Sequelize models have been updated to reflect this change. The NumberValueObject has been replaced with Uuid in the LocationId and SiteId domain entities. Additionally, the TypeOfSiteId domain entity now extends StringValueObject instead of NumberValueObject. The TypeOfSite model in Sequelize has also been updated to use a string type for the ID.