"Removed unnecessary assignment in object literals in deviceData/HP5800Almacen.js"
Explanation: The diff shows that in the file server/seeders/deviceData/HP5800Almacen.js, there were several instances where an unnecessary assignment was used in object literals. Specifically, the `locationId` property of each object in the `HP5800` array was being assigned the value of `locations.filter(location => location.name === 'Torre BNC Almacen Piso 1')[0].id` using a colon instead of an equals sign. This is unnecessary because the colon is used to define keys in an object literal, not to assign values. The correct syntax is to use an equals sign to assign the value. Therefore, the colon was replaced with an equals sign to fix this issue.