"Replace UUID with Uuid in SiteId class"
Explanation:
The diff shows a change in the `SiteId` class where the import statement for `UUID` from 'sequelize' has been replaced with `Uuid` from '../../../Shared/domain/value-object/Uuid'. This change is a modification of the type used in the class, replacing the Sequelize UUID type with an in-house Uuid type.