"Update AgenciaPage to use HardDriveCategory and adjust table headers; update TableComponent; update HomePage to use new TableComponent"
This commit message describes the changes in the code diff above. It mentions that the AgenciaPage component has been updated to use the HardDriveCategory and adjust the table headers accordingly. It also notes that the TableComponent has been updated and the HomePage component has been adjusted to use the new TableComponent.